Move CSVFieldDelimiter to a more generic place
